cleanup, fixed EditText's menu disappearing after longpress issue, fixed SeekBar's and RangeSeekBar's ripple after release/cancel, added theme for Spinner's popup

+++ carbon/src/main/java/carbon/internal/ArgbEvaluator.java
@@ -33,9 +33,9 @@ public class ArgbEvaluator implements TypeEvaluator {
         int endG = (endInt >> 8) & 0xff;
         int endB = endInt & 0xff;
 
-        return (int) ((startA + (int) (fraction * (endA - startA))) << 24) |
-                (int) ((startR + (int) (fraction * (endR - startR))) << 16) |
-                (int) ((startG + (int) (fraction * (endG - startG))) << 8) |
-                (int) ((startB + (int) (fraction * (endB - startB))));
+        return (startA + (int) (fraction * (endA - startA))) << 24 |
+                (startR + (int) (fraction * (endR - startR))) << 16 |
+                (startG + (int) (fraction * (endG - startG))) << 8 |
+                (startB + (int) (fraction * (endB - startB)));
     }
 }
